requirements.txt
streamlit 
langchain 
selenium
beautifulsoup4
lxml 
html5lib
python-dotenv
langchain-groq
dotenv
torch
transformers
sentence-transformers
faiss-cpu
numpy
Pillow
requests
asyncio
nest-asyncio
langchain_community
chromadb
langchain_google_genai
chardet
PyPDF2
python-docx
